+++ resources/assets/lib/oauth/own-clients.tsx
@@ -16,7 +16,7 @@
  *    along with osu!web.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { OwnClient } from 'oauth/own-client-component';
+import { OwnClient } from 'oauth/own-client';
 import core from 'osu-core-singleton';
 import * as React from 'react';
